[#45] Refactoring at the point where the temp test notebook runs fine. Still need to complete turning that into a proper test case.
