--- pylib/embedding_helper.py
@@ -1,426 +1,0 @@
-# SPDX-FileCopyrightText: 2023-present Oori Data <info@oori.dev>
-# SPDX-License-Identifier: Apache-2.0
-# ogbujipt.embedding_helper
-
-'''
-Helper for vector databases embeddings in such as Qdrant
-
-Vector DBs are useful when you have a lot of context to use with LLMs,
-e.g. a large document or collection of docs. One common pattern is to create
-vector indices on this text. Given an LLM prompt, the vector DB is first queried
-to find the most relevant "top k" sections of text based on the prompt,
-which are added as context in the ultimate LLM invocation.
-
-For sample code see demo/chat_pdf_streamlit_ui.py
-
-You need an LLM to turn the text into vectors for such indexing, and these
-vectors are called the embeddings. You can usually create useful embeddings
-with a less powerful (and more efficient) LLM.
-
-This module provides utilities to set up a vector DB, and use it to index
-chunks of text using a provided LLM model to create the embeddings.
-
-Other common use-cases for vector DBs in LLM applications:
-
-* Long-term LLM Memory for chat: index the entire chat history and retrieve
-the most relevant and recent N messages based on the user's new message,
-to give the LLM a chance to make its responses coherent and on-topic
-
-* Cache previous LLM interactions, saving resources by retrieving previous
-responses to similar questions without having to use the most powerful LLM
-'''
-
-import warnings
-import itertools
-from dotenv import load_dotenv
-from typing import Sequence
-import asyncpg
-from pgvector.asyncpg import register_vector
-
-# Qdrant is optional, so we'll only import it if it's available
-try:
-    from qdrant_client import QdrantClient
-    from qdrant_client.http import models
-    QDRANT_AVAILABLE = True
-except ImportError:
-    QDRANT_AVAILABLE = False
-    QdrantClient = object()  # Set up a dummy to satisfy the type hints
-    models = None
-
-# Option for running a Qdrant DB locally in memory
-MEMORY_QDRANT_CONNECTION_PARAMS = {'location': ':memory:'}
-
-load_dotenv()
-# ======================================== MOVE ME TO A NEW FILE IN THE LIBRARY ========================================
-# Generic SQL for creating a table to hold embedded documents
-CREATE_DOC_TABLE = '''\
-CREATE TABLE IF NOT EXISTS {table_name} (
-    id bigserial primary key, 
-    embedding vector({embed_dimension}),  -- embedding vector field size
-    content text NOT NULL,                -- text content of the chunk
-    permission text,                      -- permission of the chunk
-    title text,                           -- title of file
-    page_numbers integer[],               -- page number of the document that the chunk is found in
-    tags text[]                           -- tags associated with the chunk
-);\
-'''
-
-INSERT_DOC_TABLE = '''\
-INSERT INTO {table_name} (
-    embedding,
-    content,
-    permission,
-    title,
-    page_numbers,
-    tags
-) VALUES ($1, $2, $3, $4, $5, $6)
-'''
-
-SEARCH_DOC_TABLE = '''\
-SELECT 
-    1 - (embedding <=> '{search_embedding}') AS cosine_similarity,
-    title,
-    content
-FROM 
-    embeddings
-ORDER BY
-    cosine_similarity DESC
-LIMIT {limit}
-;\
-'''
-# ======================================================================================================================
-
-class PGvectorConnection:
-    def __init__(self, embedding_model, conn):
-        '''
-        Initialize a pgvector connection
-        This has limited protections against SQL injection, so be careful!
-
-        Args:
-            embedding (SentenceTransformer): SentenceTransformer object of your choice
-            https://huggingface.co/sentence-transformers
-
-            conn_params (mapping): keyword parameters for setting up a pgvector connection
-            See the main docstring (or run `help(pgvector_connection)`)
-        '''
-        # Check if the provided embedding model is a SentenceTransformer
-        if (embedding_model.__class__.__name__ == 'SentenceTransformer') and (not None):
-            self._embedding_model = embedding_model
-            self._embed_dimension = len(self._embedding_model.encode(''))
-        elif embedding_model is None:
-            self._embedding_model = None
-            self._embed_dimension = 0
-        else:
-            raise ValueError('embedding_model must be a SentenceTransformer object or None')
-
-        self.conn = conn
-
-    @classmethod
-    async def create(
-            cls,
-            embedding_model,
-            user: str,
-            password: str,
-            db_name: str,
-            host: str, 
-            port: int,
-            **conn_params
-            ) -> 'PGvectorConnection':
-        conn = await cls._set_up(user, password, db_name, host, port, **conn_params)
-        await register_vector(conn)
-        return cls(embedding_model, conn)
-
-    @staticmethod
-    async def _set_up(user, password, db_name, host, port, **conn_params):
-        try:
-            # Create a connection to the database
-            conn = await asyncpg.connect(
-                host=host,
-                port=port,
-                user=user,
-                password=password,
-                database=db_name,
-                **conn_params
-            )
-            return conn
-        except Exception as e:
-            raise ConnectionError(f"ERROR: {e}")
-
-    async def create_doc_table(self, table_name: str) -> None:
-        '''
-        Create a table to hold embedded documents
-
-        Args:
-            table_name (str): name of the table to create
-        '''
-        # Check that the connection is still alive
-        if self.conn.is_closed():
-            raise ConnectionError('Connection to database is closed')
-
-        # Create the table
-        await self.conn.execute(
-            CREATE_DOC_TABLE.format(
-                table_name=table_name,
-                embed_dimension=self._embed_dimension)
-            )
-    
-    async def insert_doc_table(
-            self, 
-            table_name: str, 
-            content: str, 
-            permission: str = "NULL", 
-            title: str = "NULL", 
-            page_numbers: list[int] = [], 
-            tags: list[str] = []
-        ) -> None:
-        '''
-        Update a table to hold embedded documents
-
-        Args:
-            table_name (str): name of the table to update
-
-            content (str): text content of the document
-
-            permission (str): permission of the document
-
-            title (str): title of the document
-
-            page_numbers (list): page number of the document that the chunk is found in
-
-            tags (list): tags associated with the document
-        '''
-        # Check that the connection is still alive
-        if self.conn.is_closed():
-            raise ConnectionError('Connection to database is closed')
-
-        # Get the embedding of the content as a PGvector compatible list
-        content_embedding = self._embedding_model.encode(content)
-        
-        await self.conn.execute(INSERT_DOC_TABLE.format(table_name=table_name), content_embedding.tolist(), content,
-                                permission, title, page_numbers, tags)
-
-    async def insert_doc_tables(
-            self,
-            table_name: str,
-            content_list: Sequence[tuple[str, str | None,  str | None, list[int], list[str]]]
-    ) -> None:
-        '''
-        Update a table to include a list of embedded documents all to be added to the same table.  Semantically
-        equivalent to calling insert_doc_tables with each tuple in content_lists as its arguments, but uses
-        executemany for efficiency
-
-        Args:
-            table_name (str): name of the table to update
-            content_list (Sequence[ .. ]): A list of tuples, each of the form: (content, permission, title,
-                                                                                page_numbers, tags)
-        '''
-        # Check that the connection is still alive
-        if self.conn.is_closed():
-            raise ConnectionError('Connection to database is closed')
-
-        await self.conn.execute(INSERT_DOC_TABLE.format(table_name=table_name), [
-            (self._embedding_model.encode(content), content, permission, title, page_numbers, tags)
-            for content, permission, title, page_numbers, tags in content_list
-        ])
-
-    async def search_doc_table(
-            self,
-            table_name: str,
-            query_string: str,
-            limit: int = 1
-            ) -> list[asyncpg.Record]:
-        '''
-        Similarity search documents using a query string
-
-        Args:
-            table_name (str): name of the table to search
-
-            query_string (str): string to compare against items in the table
-
-            k (int): maximum number of results to return (useful for top-k query)
-        Returns:
-            list[asyncpg.Record]: list of search results
-            asyncpg.Record objects are similar to dicts, but allow for attribute-style access
-        '''
-        # Check that the connection is still alive
-        if self.conn.is_closed():
-            raise ConnectionError('Connection to database is closed')
-
-        # Get the embedding of the query string as a PGvector compatible list
-        query_embedding = list(self._embedding_model.encode(query_string))
-
-        # Search the table
-        search_results = await self.conn.fetch(
-            SEARCH_DOC_TABLE.format(
-                table_name = table_name,
-                search_embedding = query_embedding,
-                limit = limit
-                )
-            )
-        
-        return search_results
-        
-
-class qdrant_collection:
-    def __init__(self, name, embedding_model, db=None,
-                 distance_function=None, **conn_params):
-        '''
-        Initialize a Qdrant client
-
-        Args:
-            name (str): of the collection
-
-            embedding (SentenceTransformer): SentenceTransformer object of your choice
-            https://huggingface.co/sentence-transformers
-
-            
-            db (optional QdrantClient): existing DB/client to use
-
-            distance_function (str): Distance function by which vectors will be compared
-
-            conn_params (mapping): keyword parameters for setting up QdrantClient
-            See the main docstring (or run `help(QdrantClient)`)
-            https://github.com/qdrant/qdrant-client/blob/master/qdrant_client/qdrant_client.py#L12
-
-        Example:
-
-        >>> from ogbujipt.text_helper import text_splitter
-        >>> from ogbujipt.embedding_helper import qdrant_collection  # pip install qdrant_client
-        >>> from sentence_transformers import SentenceTransformer  # pip install sentence_transformers
-        >>> text = 'The quick brown fox\njumps over the lazy dog,\nthen hides under a log\nwith a frog.\n'
-        >>> text += 'Should the hound wake up,\nall jumpers beware\nin a log, in a bog\nhe\'ll search everywhere.\n'
-        >>> embedding_model = SentenceTransformer('all-MiniLM-L6-v2')
-        >>> collection = qdrant_collection('my-text', embedding_model)
-        >>> chunks = text_splitter(text, chunk_size=20, chunk_overlap=4, separator='\n')
-        >>> collection.update(texts=chunks, metas=[{'seq-index': i} for (i, _) in enumerate(chunks)])
-        >>> retval = collection.search('what does the fox say?', limit=1)
-        retval
-        '''
-        self.name = name
-        self.db = db
-        # Check if the provided embedding model is a SentenceTransformer
-        if embedding_model.__class__.__name__ == 'SentenceTransformer':
-            self._embedding_model = embedding_model
-        else:
-            raise ValueError('embedding_model must be a SentenceTransformer object')
-            
-        if not self.db:
-            if not QDRANT_AVAILABLE: 
-                raise RuntimeError('Qdrant not installed, you can run `pip install qdrant-client`')
-
-            # Create a Qdrant client
-            if not conn_params:
-                conn_params = MEMORY_QDRANT_CONNECTION_PARAMS
-            self.db = QdrantClient(**conn_params)
-        self._distance_function = distance_function or models.Distance.COSINE
-        self._db_initialized = False
-
-    def _first_update_prep(self, text):
-        if text.__class__.__name__ != 'str':
-            raise ValueError('text must be a string')
-
-        # Make sure we have a vector size set; use a sample embedding if need be
-        partial_embeddings = self._embedding_model.encode(text)
-        self._vector_size = len(partial_embeddings)
-
-        # Create a collection in the Qdrant client, and configure its vectors
-        # Using REcreate_collection ensures overwrite for a clean, fresh, new collection
-        self.db.recreate_collection(
-            collection_name=self.name,
-            vectors_config=models.VectorParams(
-                size=self._vector_size, 
-                distance=self._distance_function
-                )
-            )
-
-        self._db_initialized = True
-
-    def update(self, texts, metas=None):
-        '''
-        Update/insert into a Qdrant client's collection with the some chunks of text
-
-        Args:
-            texts (List[str]): Strings to be stored and indexed. For best results these should be of similar length.
-            They'll be converted to embeddings for efficient lookup
-
-            metas (List[dict]): Optional metadata per text, stored with the text and included whenever the text is
-            retrieved via search/query
-        '''
-        if len(texts) == 0:
-            warnings.warn('Empty sequence of texts provided. No action will be taken.')
-            return
-
-        if metas is None:
-            metas = [{}]*(len(texts))
-        else:
-            if len(texts) > len(metas):
-                warnings.warn(f'More texts ({len(texts)} provided than metadata {len(metas)}).'
-                              'Extra metadata items will be ignored.')
-                metas = itertools.chain(metas, [{}]*(len(texts)-len(metas)))
-            elif len(metas) > len(texts):
-                warnings.warn(f'Fewer texts ({len(texts)} provided than metadata {len(metas)}). '
-                              'The extra text will be given empty metadata.')
-                metas = itertools.islice(metas, len(texts))
-
-        if not self._db_initialized:
-            self._first_update_prep(texts[0])
-            before_count = 0
-        else:
-            before_count = self.count()
-
-        for ix, (text, meta) in enumerate(zip(texts, metas)):
-            # Embeddings as float/vectors
-            # The inline prints actually turn into a cool progress indicator in jupyter 😁
-            embeddings = list(map(float, self._embedding_model.encode(text)))
-
-            payload = dict(_text=text, **meta)
-
-            # Upsert the embedded chunk and its payload into the collection
-            self.db.upsert(
-                collection_name=self.name,
-                points=[
-                    models.PointStruct(
-                        id=ix + before_count,  # sequential IDs
-                        vector=embeddings,
-                        payload=payload
-                        )
-                    ]
-                )
-    
-    def reset(self):
-        '''
-        Reset the Qdrant collection, deleting the collection and all its contents
-        '''
-        if not self._db_initialized:
-            raise RuntimeError('Qdrant Collection must be initialized before deleting its contents.')
-        
-        self.db.delete_collection(collection_name=self.name)
-        self._db_initialized = False
-
-    def search(self, query, **kwargs):
-        '''
-        Perform a search on this Qdrant collection
-
-        Args:
-            query (str): string to compare against items in the collection
-
-            kwargs: other args to be passed to qdrant_client.QdrantClient.search(). Common ones:
-                    limit - maximum number of results to return (useful for top-k query)
-        '''
-        if not self._db_initialized:
-            raise RuntimeError('Qdrant Collection must be initialized before searching its contents.')
-        
-        if query.__class__.__name__ != 'str':
-            raise ValueError('query must be a string')
-        embedded_query = self._embedding_model.encode(query)
-        return self.db.search(collection_name=self.name, query_vector=embedded_query, **kwargs)
-
-    def count(self):
-        '''
-        Return the count of items in this Qdrant collection
-        '''
-        if not self._db_initialized:
-            raise RuntimeError('Qdrant Collection must be initialized before counting its contents.')
-        # This ugly declaration just gets the count as an integer
-        current_count = int(str(self.db.count(self.name)).partition('=')[-1])
-        return current_count
