+++ requirements-pgvector.txt
@@ -1,0 +1,2 @@
+pgvector
+asyncpg
