+++ test/embedding/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+
+@pytest.fixture
+def CORRECT_STRING():
+    return 'And the secret thing in its heaving\nThreatens with iron mask\nThe last lighted torch of the century…'
+
+
